Fix: reverse order of triangle vertices to correct which side of triangle is the front face
Before was setting vertices in a clockwise fashion, but telling the rasterizer that the front face could be found by following the points counter clockwise. Then when the back face was culled, it was actually facing forward and just disappearing. By reversing the order of the vertices, now the front faces forward and the back face continues to be culled.